python: refactor tests and add literal datetime.date
